Move methods concerned with persistance to new file.
